refactor: create sabrmetrics.mlb.address module
